+++ src/tests/unit/utils/firebase.test.js
